1. 테마, 테마상세뷰컨트롤러를 테이블뷰컨트롤로 교체 2. 테마테이블뷰컨트롤러 앞에 네비게이션컨트롤러 임베디드
