--- WATCHA/Controller/ThemeDetailViewController.swift
@@ -1,35 +1,0 @@
-//
-//  ThemeDetailViewController.swift
-//  WATCHA
-//
-//  Created by Seo JaeHyeong on 09/04/2018.
-//  Copyright © 2018 Seo Jaehyeong. All rights reserved.
-//
-
-import UIKit
-
-class ThemeDetailViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
