--- WATCHA/Controller/ThemeViewController.swift
@@ -1,29 +1,0 @@
-//
-//  ThemeViewController.swift
-//  WATCHA
-//
-//  Created by Seo JaeHyeong on 09/04/2018.
-//  Copyright © 2018 Seo Jaehyeong. All rights reserved.
-//
-
-import UIKit
-
-class ThemeViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        <#code#>
-    }
-
-}
